+++ src/ResourceManager/Compute/Commands.Compute/ChangeLog.md
@@ -19,7 +19,9 @@
         - Additional information about change #1
 -->
 ## Current Release
+* Fixed issue with storage account type for VM with managed disk
 * Fixed issue with default resource groups not being set.
+* Fix AEM Extension cmdlets for other environments, for example Azure China
 
 ## Version 5.5.0
 * Updated to the latest version of the Azure ClientRuntime.